+++ Net_Core/BusinessLogic/BNameValuePairs.cs
@@ -56,9 +56,15 @@ public class BNameValuePairs : AbstractBusinessLogic
     /// </example>
     public BNameValuePairs(MSecurityEntity securityEntityProfile)
     {
-        if (securityEntityProfile == null) throw new ArgumentNullException(nameof(securityEntityProfile), "securityEntityProfile can not be null (Nothing in Visual Basic)");
-        m_DNameValuePairs = (INameValuePairs)ObjectFactory.Create(securityEntityProfile.DataAccessLayerAssemblyName, securityEntityProfile.DataAccessLayerNamespace, "DNameValuePairs");
-        m_DNameValuePairs.ConnectionString = securityEntityProfile.ConnectionString;
+        if (securityEntityProfile == null) throw new ArgumentNullException(nameof(securityEntityProfile), "securityEntityProfile cannot be a null reference (Nothing in Visual Basic)!");
+        if(m_DNameValuePairs == null || ConfigSettings.CentralManagement)
+        {
+            this.m_DNameValuePairs = (INameValuePairs)ObjectFactory.Create(securityEntityProfile.DataAccessLayerAssemblyName, securityEntityProfile.DataAccessLayerNamespace, "DNameValuePairs", securityEntityProfile.ConnectionString);
+            if (this.m_DNameValuePairs == null) 
+            {
+                throw new InvalidOperationException("Failed to create an instance of DNameValuePairs.");
+            }
+        }
     }
 
     /// <summary>