+++ Release_Notes.md
@@ -117,7 +117,7 @@
 - DFunctions Added a constructor that accepts connectionString so we don't have to set the properties in BFunctions.
     - BFunctions changed the constructor to pass the connectionString and the securityEntitySeqID to the DFunctions constructor
     - BFunctions now accounts for CentralManagement
-- DGroups Added a constructor that accepts connectionString so we don't have to set the properties in DGroups.
+- DGroups Added a constructor that accepts connectionString and the securityEntitySeqID so we don't have to set the properties in DGroups.
     - DGroups changed the constructor to pass the connectionString and the securityEntitySeqID to the DGroups constructor
     - DGroups now accounts for CentralManagement
 - DLogging Added a constructor that accepts connectionString so we don't have to set the properties in DLogging.
@@ -126,6 +126,10 @@
 - DMessages Added a constructor that accepts connectionString so we don't have to set the properties in DMessages.
     - BMessages changed the constructor to pass the connectionString to the DMessages constructor
     - BMessages now accounts for CentralManagement
+- DNameValuePairs Added a constructor that accepts connectionString and the securityEntitySeqID so we don't have to set the properties in BNameValuePairs.
+    - BNameValuePairs changed the constructor to pass the connectionString and the securityEntitySeqID to the DNameValuePairs constructor
+    - BNameValuePairs now accounts for CentralManagement
+
 
 ### Known Bugs
 - There is an issue where the refresh token is being revoked when it shouldn't be causing a logout to occure